Split mastercore_dex into mastercore_dex, mastercore_mdex
