+++ src/mastercore_rpc.cpp
@@ -5,6 +5,7 @@
 #include "mastercore.h"
 #include "mastercore_convert.h"
 #include "mastercore_dex.h"
+#include "mastercore_mdex.h"
 #include "mastercore_errors.h"
 #include "mastercore_log.h"
 #include "mastercore_parse_string.h"
