+++ src/mastercore_tx.cpp
@@ -5,11 +5,13 @@
 #include "mastercore.h"
 #include "mastercore_convert.h"
 #include "mastercore_dex.h"
+#include "mastercore_mdex.h"
 #include "mastercore_log.h"
 #include "mastercore_sp.h"
 
 #include "alert.h"
 #include "amount.h"
+#include "main.h"
 #include "utiltime.h"
 
 #include <boost/algorithm/string.hpp>
